# ProyectoFinalAvanzada/src/main/gen/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: API para reportar incidentes ciudadanos en tiempo real.
  title: Plataforma de Reportes de Seguridad
  version: 1.0.0
servers:
- description: Servidor de producción
  url: https://api.reportes-ciudad.com/v1
- description: Servidor local para desarrollo
  url: http://localhost:8080
paths:
...
  /admin/perfil:
    put:
      description: "Permite a un administrador modificar su nombre, correo y contraseñ\
        a."
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminEditar'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_admin_perfil_put_200_response'
          description: Perfil de administrador actualizado con éxito.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_reportes__id__put_400_response'
          description: Datos inválidos.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_admin_perfil_put_401_response'
          description: No autorizado.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_admin_perfil_put_404_response'
          description: Administrador no encontrado.
      summary: Editar perfil del administrador
      tags:
      - admin
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: admin
  /admin/categorias:
    post:
      description: Permite al administrador crear una categoria de reporte
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Categoria'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_admin_categorias_post_201_response'
          description: Reporte creado con éxito
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_admin_categorias_post_400_response'
          description: Datos inválidos
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_reportes_post_500_response'
          description: Error interno del servidor
      summary: Crear una categoria
      tags:
      - admin
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: admin
  /admin/categorias/{id}:
    delete:
      description: Permite eliminar una categoría por su ID.
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_admin_categorias__id__delete_200_response'
          description: Categoría eliminada con éxito.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_admin_categorias__id__delete_400_response'
          description: No se puede eliminar la categoría.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_admin_categorias__id__put_404_response'
          description: Categoría no encontrada.
      summary: Eliminar una categoría
      tags:
      - admin
      x-accepts:
      - application/json
      x-tags:
      - tag: admin
    put:
      description: Permite modificar el nombre y la descripción de una categoría.
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoriaEditar'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_admin_categorias__id__put_200_response'
          description: Categoría actualizada con éxito.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_reportes__id__put_400_response'
          description: Datos inválidos.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_admin_categorias__id__put_404_response'
          description: Categoría no encontrada.
      summary: Editar una categoría
      tags:
      - admin
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: admin
...
components:
  schemas:
...
    Categoria:
      example:
        descripcion: "Reportes de accidentes de tránsito, problemas de salud en la\
          \ vía pública, etc."
        id: id
        nombre: Emergencias Médicas
      properties:
        id:
          type: string
        nombre:
          example: Emergencias Médicas
          type: string
        descripcion:
          example: "Reportes de accidentes de tránsito, problemas de salud en la ví\
            a pública, etc."
          type: string
      required:
      - descripcion
      - nombre
      type: object
    CategoriaEditar:
      example:
        descripcion: "Reportes de terremotos, incendios, y otras emergencias."
        nombre: Emergencias y Desastres
      properties:
        nombre:
          example: Emergencias y Desastres
          type: string
        descripcion:
          example: "Reportes de terremotos, incendios, y otras emergencias."
          type: string
      type: object
...
    AdminEditar:
      example:
        correo: admin@empresa.com
        nombre: Administrador Principal
        contraseña: NuevaClaveSegura123
      properties:
        nombre:
          example: Administrador Principal
          type: string
        correo:
          example: admin@empresa.com
          format: email
          type: string
        contraseña:
          example: NuevaClaveSegura123
          format: password
          type: string
      required:
      - contraseña
      - correo
      - nombre
      type: object
...
    _reportes_post_500_response:
      example:
        error: Error inesperado en el servidor. Inténtelo más tarde.
      properties:
        error:
          example: Error inesperado en el servidor. Inténtelo más tarde.
          type: string
      type: object
...
    _reportes__id__put_400_response:
      example:
        error: Los datos proporcionados no son válidos.
      properties:
        error:
          example: Los datos proporcionados no son válidos.
          type: string
      type: object
...
    _admin_perfil_put_200_response:
      example:
        mensaje: Perfil de administrador actualizado correctamente.
      properties:
        mensaje:
          example: Perfil de administrador actualizado correctamente.
          type: string
      type: object
    _admin_perfil_put_401_response:
      example:
        error: No tienes permisos para editar este perfil.
      properties:
        error:
          example: No tienes permisos para editar este perfil.
          type: string
      type: object
    _admin_perfil_put_404_response:
      example:
        error: Administrador no encontrado.
      properties:
        error:
          example: Administrador no encontrado.
          type: string
      type: object
    _admin_categorias_post_201_response:
      example:
        id: id
        mensaje: Categoria creada con exito
      properties:
        id:
          type: string
        mensaje:
          example: Categoria creada con exito
          type: string
      type: object
    _admin_categorias_post_400_response:
      example:
        error: Faltan datos necesarios para la categoria
      properties:
        error:
          example: Faltan datos necesarios para la categoria
          type: string
      type: object
    _admin_categorias__id__put_200_response:
      example:
        mensaje: Categoría editada correctamente.
      properties:
        mensaje:
          example: Categoría editada correctamente.
          type: string
      type: object
    _admin_categorias__id__put_404_response:
      example:
        error: Categoría no encontrada.
      properties:
        error:
          example: Categoría no encontrada.
          type: string
      type: object
    _admin_categorias__id__delete_200_response:
      example:
        mensaje: Categoría eliminada correctamente.
      properties:
        mensaje:
          example: Categoría eliminada correctamente.
          type: string
      type: object
    _admin_categorias__id__delete_400_response:
      example:
        error: No se puede eliminar la categoría porque tiene reportes asociados.
      properties:
        error:
          example: No se puede eliminar la categoría porque tiene reportes asociados.
          type: string
      type: object
...